refactor: post abstraction
